backend/requirements.txt
fastapi[standard]
python-jose[cryptography]
passlib
bcrypt==4.0.1
python-multipart
sqlalchemy
pydantic[email]
pydantic-settings
uvicorn[standard]